Feature: Create interface test for login #159
